# Decrease allocations in TextFindEngine/SelectionWordBreaker, improve performance