--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Documents/SelectionWordBreaker.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Documents/SelectionWordBreaker.cs
@@ -73,7 +73,7 @@ namespace System.Windows.Documents
                 }
             }
 
-            ReadOnlySpan<UInt16> charType3 = stackalloc UInt16[2];
+            Span<UInt16> charType3 = stackalloc UInt16[2];
             ReadOnlySpan<char> sourceChars = [text[position - 1], text[position]];
 
             SafeNativeMethods.GetStringTypeEx(0 /* ignored */, SafeNativeMethods.CT_CTYPE3, sourceChars, charType3);
@@ -203,7 +203,7 @@ namespace System.Windows.Documents
         private static CharClass[] GetClasses(char[] text)
         {
             CharClass[] classes = new CharClass[text.Length];
-            ReadOnlySpan<UInt16> charType1 = stackalloc UInt16[1];
+            UInt16 charType1 = UInt16.MinValue;
 
             for (int i = 0; i < text.Length; i++)
             {
@@ -228,11 +228,11 @@ namespace System.Windows.Documents
                 }
                 else
                 {
-                    SafeNativeMethods.GetStringTypeEx(0 /* ignored */, SafeNativeMethods.CT_CTYPE1, new(in ch), charType1);
+                    SafeNativeMethods.GetStringTypeEx(0 /* ignored */, SafeNativeMethods.CT_CTYPE1, [ch], new Span<UInt16>(ref charType1));
 
-                    if ((charType1[0] & SafeNativeMethods.C1_SPACE) != 0)
+                    if ((charType1 & SafeNativeMethods.C1_SPACE) != 0)
                     {
-                        if ((charType1[0] & SafeNativeMethods.C1_BLANK) != 0)
+                        if ((charType1 & SafeNativeMethods.C1_BLANK) != 0)
                         {
                             classification = CharClass.Blank | CharClass.WBF_ISWHITE;
                         }
@@ -241,7 +241,7 @@ namespace System.Windows.Documents
                             classification = CharClass.WhiteSpace | CharClass.WBF_ISWHITE;
                         }
                     }
-                    else if ((charType1[0] & SafeNativeMethods.C1_PUNCT) != 0 && !IsDiacriticOrKashida(ch))
+                    else if ((charType1 & SafeNativeMethods.C1_PUNCT) != 0 && !IsDiacriticOrKashida(ch))
                     {
                         classification = CharClass.Punctuation;
                     }
@@ -260,11 +260,11 @@ namespace System.Windows.Documents
         // Returns true if a char is a non-spacing diacritic or kashida.
         private static bool IsDiacriticOrKashida(char ch)
         {
-            ReadOnlySpan<UInt16> charType3 = stackalloc UInt16[1];
+            UInt16 charType3 = UInt16.MinValue;
 
-            SafeNativeMethods.GetStringTypeEx(0 /* ignored */, SafeNativeMethods.CT_CTYPE3, new(in ch), charType3);
+            SafeNativeMethods.GetStringTypeEx(0 /* ignored */, SafeNativeMethods.CT_CTYPE3, [ch], new Span<UInt16>(ref charType3));
 
-            return (charType3[0] & (SafeNativeMethods.C3_DIACRITIC | SafeNativeMethods.C3_NONSPACING | SafeNativeMethods.C3_VOWELMARK | SafeNativeMethods.C3_KASHIDA)) != 0;
+            return (charType3 & (SafeNativeMethods.C3_DIACRITIC | SafeNativeMethods.C3_NONSPACING | SafeNativeMethods.C3_VOWELMARK | SafeNativeMethods.C3_KASHIDA)) != 0;
         }
 
         // Returns true if a character falls within a specified code point range.