## Description

Removes unnecessary allocations of temporary heap arrays when calling `GetStringTypeEx`, using stack-allocated space instead.
Also removes some unnecessary string copies in `TextFindEngine.Find(...)` method.

### Base replication of `Find` method
| Method   | findPattern          | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |--------------------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original |  this(...)e it  [92] |  58.53 ns |   1.161 ns |    1.193 ns | 0.0200 |         461 B |         336 B |
| PR__EDIT |  this(...)e it  [92] |  28.89 ns |   0.213 ns |    0.199 ns |      - |         350 B |             - |

<details><summary>Benchmark code</summary>

```csharp
[Benchmark]
[Arguments(" this is a very long string, and yet its gonna be even longer string now, hope u can see it ")]
public UInt16 Original(string findPattern)
{
    UInt16[] findPatternStartCharType1 = new UInt16[1];
    UInt16[] findPatternEndCharType1 = new UInt16[1];
    char[] charFindPattern = findPattern.ToCharArray();

    // Get the character type for the start/end character of the find pattern.
    SafeNativeMethods.GetStringTypeEx(0 /* ignored */, SafeNativeMethods.CT_CTYPE1, new char[] { charFindPattern[0] }, 1, findPatternStartCharType1);
    SafeNativeMethods.GetStringTypeEx(0 /* ignored */, SafeNativeMethods.CT_CTYPE1, new char[] { charFindPattern[charFindPattern.Length - 1] }, 1, findPatternEndCharType1);

    return (ushort)(findPatternStartCharType1[0] + findPatternEndCharType1[0]);
}

[Benchmark]
[Arguments(" this is a very long string, and yet its gonna be even longer string now, hope u can see it ")]
public UInt16 PR__EDIT(string findPattern)
{
    ReadOnlySpan<UInt16> startCharType1 = stackalloc UInt16[1];
    ReadOnlySpan<UInt16> endCharType1 = stackalloc UInt16[1];
    ReadOnlySpan<char> startEndChars = [findPattern[0], findPattern[findPattern.Length - 1]];

    // Get the character type for the start/end character of the find pattern.
    SafeNativeMethods.GetStringTypeEx2(0 /* ignored */, SafeNativeMethods.CT_CTYPE1, new(in startEndChars[0]), startCharType1);
    SafeNativeMethods.GetStringTypeEx2(0 /* ignored */, SafeNativeMethods.CT_CTYPE1, new(in startEndChars[1]), endCharType1);

    return (ushort)(startCharType1[0] + endCharType1[0]);
}
```

</details>

## Customer Impact

Decreased allocations, improved performance.

## Regression

No.

## Testing

Local build, verifying the outputs of both functions.

## Risk

Low.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9553)