@himgoyalmicro 2025-01-21
@h3xds1nz the PR is ready to be merged. Just a small request can you also update https://github.com/dotnet/wpf/blob/main/src/Microsoft.DotNet.Wpf/tests/IntegrationTests/Common/DRT/TestServices/MS/Win32/SafeNativeMethodsOther.cs to have the parity between the two files.