@h3xds1nz 2025-01-21
@himgoyalmicro Yeah, lemme take a look.